modal integration home page fix
