+++ src/favorite.html
@@ -51,9 +51,10 @@
         </div> -->
         </div>
       </div>
+      <load src="./partials/scroll-up.html" />
     </main>
-    <load src="./partials/scroll-up.html" />
     <load src="./partials/footer.html" />
+    <load src="./partials/modal.html" />
 
     <script type="module" src="./js/main.js"></script>
     <script type="module" src="./js/favorite.js"></script>
