+++ src/js/favorite.js
@@ -1,5 +1,6 @@
 import { renderExercises } from './exercises.js';
-// import { LoadListenersForOpenModal } from './modal.js';
+import { LoadListenersForOpenModal } from './modal.js';
+import { GetFavorites } from './utils/local-storage.js';
 
 document.addEventListener('DOMContentLoaded', function () {
   const noFavoritesMessage = document.getElementById('no-favorites');
@@ -20,7 +21,7 @@ document.addEventListener('DOMContentLoaded', function () {
   // localStorage.setItem('favorites', JSON.stringify(sampleFavoritesData.results));
   // localStorage.removeItem('favorites');
 
-  let favorites = JSON.parse(localStorage.getItem('favorites')) || [];
+  let favorites = GetFavorites();
 
   function displayFavorites() {
     cardList.innerHTML = '';
@@ -29,7 +30,8 @@ document.addEventListener('DOMContentLoaded', function () {
       noFavoritesMessage.style.display = 'block';
     } else {
       noFavoritesMessage.style.display = 'none';
-      cardList = renderExercises({ results: favorites }, true);
+      renderExercises({ results: favorites }, true);
+      LoadListenersForOpenModal();
     }
   }
