+++ src/js/filters.js
@@ -1,7 +1,7 @@
 import { request } from './services/api-service';
 import { currentResolution } from './utils/utils.js';
 import { renderExercises } from './exercises.js';
-import { LoadModalData } from './modal.js';
+import { LoadListenersForOpenModal } from './modal.js';
 
 const filterButtons = document.querySelectorAll('.filter-button');
 const searchForm = document.querySelector('#search-form');
@@ -245,8 +245,8 @@ const handleSearch = async () => {
 
   const exercises = await request('exercises', searchParams);
 
-    renderExercises(exercises, false);
-    LoadListenersForOpenModal();
+  renderExercises(exercises, false);
+  LoadListenersForOpenModal();
 
   searchInput.value = '';
   toggleClearButton();
@@ -255,13 +255,3 @@ const handleSearch = async () => {
 searchInput.addEventListener('input', () => {
   toggleClearButton();
 });
-
-function LoadListenersForOpenModal() {
-  const btnOpenModal = document.querySelectorAll("[data-modal-open]");
-  for (let i = 0; i < btnOpenModal.length; i++) {
-    // console.log(btnOpenModal[i]);
-    btnOpenModal[i].addEventListener("click", (event) => {
-      const btn = event.currentTarget;
-      LoadModalData(btn.value);
-    });  
-  }}
